// src/models/Common.namespace.ts
export namespace CommonNamespace {
  export enum LevelsEnum {
    trash = 'trash',
    usual = 'usual',
    good = 'good',
    excellent = 'excellent',
    heroic = 'heroic',
    legendary = 'legendary',
    mythical = 'mythical',
  }

  export enum MaterialsEnum {
    wood = 'wood',
    iron = 'iron',
    steel = 'steel',
    gold = 'gold',
  }
}